# src/internal/openapi/ogen/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package ogen

import (
	"fmt"
)

func (s *ErrorResponseStatusCode) Error() string {
	return fmt.Sprintf("code %d: %+v", s.StatusCode, s.Response)
}

// Ref: #/components/schemas/ErrorResponse
type ErrorResponse struct {
	Error string `json:"error"`
}

// GetError returns the value of Error.
func (s *ErrorResponse) GetError() string {
	return s.Error
}

// SetError sets the value of Error.
func (s *ErrorResponse) SetError(val string) {
	s.Error = val
}

func (*ErrorResponse) createSubjectRelationRes() {}
func (*ErrorResponse) createSubjectRes()         {}

// ErrorResponseStatusCode wraps ErrorResponse with StatusCode.
type ErrorResponseStatusCode struct {
	StatusCode int
	Response   ErrorResponse
}

// GetStatusCode returns the value of StatusCode.
func (s *ErrorResponseStatusCode) GetStatusCode() int {
	return s.StatusCode
}

// GetResponse returns the value of Response.
func (s *ErrorResponseStatusCode) GetResponse() ErrorResponse {
	return s.Response
}

// SetStatusCode sets the value of StatusCode.
func (s *ErrorResponseStatusCode) SetStatusCode(val int) {
	s.StatusCode = val
}

// SetResponse sets the value of Response.
func (s *ErrorResponseStatusCode) SetResponse(val ErrorResponse) {
	s.Response = val
}

// GetSubjectByNameNotFound is response for GetSubjectByName operation.
type GetSubjectByNameNotFound struct{}

func (*GetSubjectByNameNotFound) getSubjectByNameRes() {}

// GetSubjectNotFound is response for GetSubject operation.
type GetSubjectNotFound struct{}

func (*GetSubjectNotFound) getSubjectRes() {}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/components/schemas/PingResponse
type PingResponse struct {
	Message string `json:"message"`
}

// GetMessage returns the value of Message.
func (s *PingResponse) GetMessage() string {
	return s.Message
}

// SetMessage sets the value of Message.
func (s *PingResponse) SetMessage(val string) {
	s.Message = val
}

// Ref: #/components/schemas/Subject
type Subject struct {
	ID                    string   `json:"id"`
	Name                  string   `json:"name"`
	RelatedToSubjectIds   []string `json:"related_to_subject_ids"`
	RelatedFromSubjectIds []string `json:"related_from_subject_ids"`
}

// GetID returns the value of ID.
func (s *Subject) GetID() string {
	return s.ID
}

// GetName returns the value of Name.
func (s *Subject) GetName() string {
	return s.Name
}

// GetRelatedToSubjectIds returns the value of RelatedToSubjectIds.
func (s *Subject) GetRelatedToSubjectIds() []string {
	return s.RelatedToSubjectIds
}

// GetRelatedFromSubjectIds returns the value of RelatedFromSubjectIds.
func (s *Subject) GetRelatedFromSubjectIds() []string {
	return s.RelatedFromSubjectIds
}

// SetID sets the value of ID.
func (s *Subject) SetID(val string) {
	s.ID = val
}

// SetName sets the value of Name.
func (s *Subject) SetName(val string) {
	s.Name = val
}

// SetRelatedToSubjectIds sets the value of RelatedToSubjectIds.
func (s *Subject) SetRelatedToSubjectIds(val []string) {
	s.RelatedToSubjectIds = val
}

// SetRelatedFromSubjectIds sets the value of RelatedFromSubjectIds.
func (s *Subject) SetRelatedFromSubjectIds(val []string) {
	s.RelatedFromSubjectIds = val
}

// Ref: #/components/schemas/SubjectCreateRequest
type SubjectCreateRequest struct {
	Data SubjectIn `json:"data"`
}

// GetData returns the value of Data.
func (s *SubjectCreateRequest) GetData() SubjectIn {
	return s.Data
}

// SetData sets the value of Data.
func (s *SubjectCreateRequest) SetData(val SubjectIn) {
	s.Data = val
}

// Ref: #/components/schemas/SubjectCreateResponse
type SubjectCreateResponse struct {
	Data Subject `json:"data"`
}

// GetData returns the value of Data.
func (s *SubjectCreateResponse) GetData() Subject {
	return s.Data
}

// SetData sets the value of Data.
func (s *SubjectCreateResponse) SetData(val Subject) {
	s.Data = val
}

func (*SubjectCreateResponse) createSubjectRes() {}

// Ref: #/components/schemas/SubjectGetResponse
type SubjectGetResponse struct {
	Data Subject `json:"data"`
}

// GetData returns the value of Data.
func (s *SubjectGetResponse) GetData() Subject {
	return s.Data
}

// SetData sets the value of Data.
func (s *SubjectGetResponse) SetData(val Subject) {
	s.Data = val
}

func (*SubjectGetResponse) getSubjectByNameRes() {}
func (*SubjectGetResponse) getSubjectRes()       {}

// Ref: #/components/schemas/SubjectIn
type SubjectIn struct {
	Name string `json:"name"`
}

// GetName returns the value of Name.
func (s *SubjectIn) GetName() string {
	return s.Name
}

// SetName sets the value of Name.
func (s *SubjectIn) SetName(val string) {
	s.Name = val
}

// Ref: #/components/schemas/SubjectRelation
type SubjectRelation struct {
	SubjectID        string `json:"subject_id"`
	RelatedSubjectID string `json:"related_subject_id"`
}

// GetSubjectID returns the value of SubjectID.
func (s *SubjectRelation) GetSubjectID() string {
	return s.SubjectID
}

// GetRelatedSubjectID returns the value of RelatedSubjectID.
func (s *SubjectRelation) GetRelatedSubjectID() string {
	return s.RelatedSubjectID
}

// SetSubjectID sets the value of SubjectID.
func (s *SubjectRelation) SetSubjectID(val string) {
	s.SubjectID = val
}

// SetRelatedSubjectID sets the value of RelatedSubjectID.
func (s *SubjectRelation) SetRelatedSubjectID(val string) {
	s.RelatedSubjectID = val
}

// Ref: #/components/schemas/SubjectRelationCreateRequest
type SubjectRelationCreateRequest struct {
	Data SubjectRelationIn `json:"data"`
}

// GetData returns the value of Data.
func (s *SubjectRelationCreateRequest) GetData() SubjectRelationIn {
	return s.Data
}

// SetData sets the value of Data.
func (s *SubjectRelationCreateRequest) SetData(val SubjectRelationIn) {
	s.Data = val
}

// Ref: #/components/schemas/SubjectRelationCreateResponse
type SubjectRelationCreateResponse struct {
	Data SubjectRelation `json:"data"`
}

// GetData returns the value of Data.
func (s *SubjectRelationCreateResponse) GetData() SubjectRelation {
	return s.Data
}

// SetData sets the value of Data.
func (s *SubjectRelationCreateResponse) SetData(val SubjectRelation) {
	s.Data = val
}

func (*SubjectRelationCreateResponse) createSubjectRelationRes() {}

// Ref: #/components/schemas/SubjectRelationIn
type SubjectRelationIn struct {
	SubjectID        string `json:"subject_id"`
	RelatedSubjectID string `json:"related_subject_id"`
}

// GetSubjectID returns the value of SubjectID.
func (s *SubjectRelationIn) GetSubjectID() string {
	return s.SubjectID
}

// GetRelatedSubjectID returns the value of RelatedSubjectID.
func (s *SubjectRelationIn) GetRelatedSubjectID() string {
	return s.RelatedSubjectID
}

// SetSubjectID sets the value of SubjectID.
func (s *SubjectRelationIn) SetSubjectID(val string) {
	s.SubjectID = val
}

// SetRelatedSubjectID sets the value of RelatedSubjectID.
func (s *SubjectRelationIn) SetRelatedSubjectID(val string) {
	s.RelatedSubjectID = val
}

// TruncateTablesOK is response for TruncateTables operation.
type TruncateTablesOK struct{}