# src/openapi-schema.yaml
openapi: 3.0.3
info:
  title: Items API
  description: An example CRUD API for Items
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /subjects:
    post:
      operationId: createSubject
      description: Creates Subject.
      requestBody:
        description: Subject to create.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectCreateRequest'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectCreateResponse'
        '409':
          description: Conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        'default':
          description: Unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: getSubjectByName
      description: Returns subject by name.
      parameters:
        - name: name
          in: query
          description: Lookup subject by name.
          required: false
          schema:
            type: string
            example: "foo"
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectGetResponse'
        '404':
          description: Not found.
        'default':
          description: Unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /subjects/{subjectUuid}:
    get:
      operationId: getSubject
      description: Returns a single Subject by uuid.
      parameters:
        - name: subjectUuid
          in: path
          description: Subject UUID
          required: true
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectGetResponse'
        '404':
          description: Not found.
        'default':
          description: Unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /subject-relations:
    post:
      operationId: createSubjectRelation
      description: Creates Subject Relation.
      requestBody:
        description: Subject Relation to create.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectRelationCreateRequest'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectRelationCreateResponse'
        '409':
          description: Conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        'default':
          description: Unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /truncate-tables:
    post:
      operationId: truncateTables
      description: Truncates all tables.
      responses:
        '200':
          description: OK.
        'default':
          description: Unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ping:
    get:
      operationId: ping
      description: Check if the service is running.
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PingResponse'
        'default':
          description: Unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    # Subject schemas

    SubjectIn:
      type: object
      properties:
        name:
          type: string
          example: foo
      required:
        - name

    Subject:
      type: object
      properties:
        id:
          type: string
          format: string
          example: 4:16987f9f-8033-4cd6-a7c6-94493f1f20ba:12
        name:
          type: string
          example: foo
        related_to_subject_ids:
          type: array
          items:
            type: string
            format: string
          example: ["4:16987f9f-8033-4cd6-a7c6-94493f1f20ba:12", "4:16987f9f-8033-4cd6-a7c6-94493f1f20ba:12"]
        related_from_subject_ids:
          type: array
          items:
            type: string
            format: string
          example: ["4:16987f9f-8033-4cd6-a7c6-94493f1f20ba:12", "4:16987f9f-8033-4cd6-a7c6-94493f1f20ba:12"]
      required:
        - id
        - name
        - related_to_subject_ids
        - related_from_subject_ids

    SubjectRelationIn:
      type: object
      properties:
        subject_id:
          type: string
          format: string
          example: "4:16987f9f-8033-4cd6-a7c6-94493f1f20ba:12"
        related_subject_id:
          type: string
          format: string
          example: "4:16987f9f-8033-4cd6-a7c6-94493f1f20ba:12"
      required:
        - subject_id
        - related_subject_id

    SubjectRelation:
      type: object
      properties:
        subject_id:
          type: string
          format: string
          example: "4:16987f9f-8033-4cd6-a7c6-94493f1f20ba:12"
        related_subject_id:
          type: string
          format: string
          example: "4:16987f9f-8033-4cd6-a7c6-94493f1f20ba:12"
      required:
        - subject_id
        - related_subject_id

    SubjectCreateRequest:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/SubjectIn'
      required:
        - data

    SubjectCreateResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Subject'
      required:
        - data

    SubjectGetResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Subject'
      required:
        - data

    SubjectRelationCreateRequest:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/SubjectRelationIn'
      required:
        - data

    SubjectRelationCreateResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/SubjectRelation'
      required:
        - data

    # Ping schemas

    PingResponse:
      type: object
      properties:
        message:
          type: string
          example: "pong"
      required:
        - message

    # Error schemas

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error